+++ doomgeneric/w_file_stdc.c
@@ -66,7 +66,7 @@ static void W_StdC_CloseFile(wad_file_t *wad)
 // Read data from the specified position in the file into the 
 // provided buffer.  Returns the number of bytes read.
 
-size_t W_StdC_Read(wad_file_t *wad, unsigned int offset,
+int W_StdC_Read(wad_file_t *wad, unsigned int offset,
                    void *buffer, size_t buffer_len)
 {
     stdc_wad_file_t *stdc_wad;
@@ -82,7 +82,7 @@ size_t W_StdC_Read(wad_file_t *wad, unsigned int offset,
 
     result = fread(buffer, 1, buffer_len, stdc_wad->fstream);
 
-    return result;
+    return (int)result;
 }
